Refactor mock webserver into a fixture we can reuse.
