+++ tests/customlists/conftest.py
@@ -1,0 +1,3 @@
+pytest_plugins = [
+    "tests.fixtures.webserver",
+]
